% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{load_raster_LSAT}
\alias{load_raster_LSAT}
\title{load all raster files stored in the RSecologyData package}
\usage{
load_raster_LSAT(overwrite = TRUE)

load_raster_LSAT(overwrite = TRUE)
}
\description{
All files are spatial raster sets (bricks or single raster)

All files are spatial raster sets (bricks or single raster)
}
\details{
the files are

lsatp224r63_1988: Landsat ... (date, 1988; path 224 row 63, projection UTM north WGS84)

lsatp224r63_2005: Landsat ... (date, 2005; path 224 row 63, projection UTM north WGS84)

the files are

lsatp224r63_1988: Landsat ... (date, 1988; path 224 row 63, projection UTM north WGS84)

lsatp224r63_2005: Landsat ... (date, 2005; path 224 row 63, projection UTM north WGS84)
}
\examples{
load_raster_LSAT()
plotRGB(lsat224r63_2005,1,2,3,stretch="lin")
load_vector()
plot(pa,add=T)
plot(abspres_poi,col="yellow",add=T)
load_raster_LSAT()
plotRGB(lsat224r63_2005,1,2,3,stretch="lin")
load_vector()
plot(pa,add=T)
plot(abspres_poi,col="yellow",add=T)
}
\author{
Martin Wegmann

Martin Wegmann
}
\references{
\href{http://remote-sensing-biodiversity.org/rsecology-book}.

\href{http://remote-sensing-biodiversity.org/rsecology-book}.
}
\seealso{
load_vector()

load_vector()
}
\keyword{raster}

