% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{load_raster_MODIS}
\alias{load_raster_MODIS}
\title{load MODIS raster files stored in the RSecologyData package}
\usage{
load_raster_MODIS(overwrite = TRUE)
}
\description{
All files are spatial raster sets (bricks or single raster)
}
\details{
MODIS files

"MCD43A42005_193_201.005.b1-7.UTMnorth": MODIS data  ... (July 2005, h 13+12, v 9, mosaiced, projection UTM north WGS84)

lsatp224r63_2005: Landsat ... (date, 2005; path 224 row 63, projection UTM north WGS84)
}
\examples{
load_raster_MODIS()
plotRGB(lsat224r63_2005,1,2,3,stretch="lin")
}
\author{
Martin Wegmann
}
\references{
\href{http://remote-sensing-biodiversity.org/rsecology-book}.
}
\seealso{
load_raster_LSAT(), load_vector()
}
\keyword{raster}

