% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadvector}
\alias{loadvector}
\title{load all vector files for RS eco}
\usage{
loadvector(overwrite = TRUE)
}
\description{
kurzbeschreibung
}
\details{
details
}
\examples{
loadvector()
plot(lsat_extent_LL)
}
\author{
Martin Wegmann
}
\seealso{
loadraster()
}
\keyword{data}
\keyword{frames}
\keyword{shape}
\keyword{spatial}
\keyword{vector,}

