% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{lsat_extent_LL}
\alias{lsat_extent_LL}
\title{data: vector file with the extent of the Landsat scene path 224 r 63 - SpatialPolygonsDataFrame}
\description{
Contains a polygon of the area of the Landsat scene with path 224 and row 63 of the WRS2 system.
}
\examples{
plot(lsat_extent_LL)
}
\references{
\href{http://landsat.usgs.gov/tools_wrs-2_shapefile.php}{WRS-1 Shape files
courtesy of the U.S. Geological Survey}.
}
\seealso{
wrs2_asc_desc
}

