% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{wrs2_asc_desc}
\alias{wrs2_asc_desc}
\title{Worldwide Reference System - 2 (WRS2) SpatialPolygonsDataFrame}
\description{
Contains polygons of the area of each path and row for the ascending and
descending nodes of the WRS2 system.
}
\references{
\href{http://landsat.usgs.gov/tools_wrs-2_shapefile.php}{WRS-2 Shape files
courtesy of the U.S. Geological Survey}.
}
\seealso{
wrs1_asc_desc
}

