% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load_raster}
\alias{load_raster}
\title{load all raster files stored in the RSecologyData package}
\usage{
load_raster(overwrite = TRUE)
}
\description{
All files are spatial raster sets (bricks or single raster)
}
\details{
the files are
lsatp224r63_2005: Landsat ... (date, 2005; path 224 row 63, projection UTM north WGS84)
}
\examples{
load_vector()
plot(lsat_extent_LL)
##
plot(pa)
plot(abspres_poi,col="yellow",add=T)
}
\author{
Martin Wegmann
}
\references{
\href{http://remote-sensing-biodiversity.org/rsecology-book}.
}
\seealso{
load_vector()
}
\keyword{raster}

