% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{load_vector}
\alias{load_vector}
\alias{load_vector_MODIS}
\title{load all vector files stored in the RSecologyData package}
\usage{
load_vector(overwrite = TRUE)

load_vector_MODIS(overwrite = TRUE)
}
\description{
All files are spatial vector data sets (Polygons, Lines or Points) providing information
such as border of protected areas, field sampling sites or roads.

Vector files related to MODIS e.g. the MODIS h-v grid.
}
\details{
the files are

"lsat_extent_LL": landsat extent for path 224 r63 (geographic coordinates lat long)

"lsat_extent": landsat extent for path 224 r63 (projection UTM north WGS84)

"study_area": rectangle of the study area (projection UTM north WGS84)

"pa": protected areas of the area covered by lsat_extent (projection UTM north WGS84)

"roads": roads of the area covered by lsat_extent (projection UTM north WGS84)

"training_vector_data": polygons with land cover classes for land cover classification (projection UTM north WGS84)

"change_classes": polygons with land cover change classes for change detection (projection UTM north WGS84)

"abspres_poi": point data with absence and presence data (0-1) (projection UTM north WGS84)

"pres_poi": point data with presence data only (projection UTM north WGS84)

"count_poi": point data with count data (whole numbers) (projection UTM north WGS84)

"trend_poi": point data with trend data (negative and positive values) (projection UTM north WGS84)

The vector files are related to MODIS and provide information in sinusoidal projection on country boundaries or the MODIS grid.
All files are spatial vector data sets (Polygons, Lines or Points).
Data compiled by Markus Neteler (FMACH) using GRASS (CC-BY-SA: http://gis.cri.fmach.it/modis-sinusoidal-gis-files/)

"MODIS_tiles_sinu": MODIS tiles (h-v) in sinusoidal projection

"MODIS_bnd_sinu": country boundaries in sinusoidal projection

"MODIS_world_sinu": the borders of the map in sinusoidal projection
}
\examples{
load_vector()
## plotting the data in geographic coordinates
plot(lsat_extent_LL)

## plotting the data in UTM north projection
plot(pa)
plot(abspres_poi,col="yellow",add=T)
load_vector_MODIS()
## plotting the world and country boundaries in sinusoidal projection
plot(MODIS_world_sinu)
plot(MODIS_bnd_sinu, add=T)

## adding the MODIS grid
plot(MODIS_tiles_sinu,add=T)
}
\author{
Martin Wegmann

Martin Wegmann
}
\references{
\href{http://remote-sensing-biodiversity.org/rsecology-book}.

\href{http://remote-sensing-biodiversity.org/rsecology-book}.
}
\seealso{
load_raster()

load_vector(), load_raster(), load_raster_MODIS()
}
\keyword{MODIS}
\keyword{data}
\keyword{frames}
\keyword{frames,}
\keyword{shape}
\keyword{spatial}
\keyword{vector,}

