% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{load_vector}
\alias{load_vector}
\title{load all vector files stored in the RSecologyData package}
\usage{
load_vector(overwrite = TRUE)
}
\description{
All files are spatial vector data sets (Polygons, Lines or Points) providing information
such as border of protected areas, field sampling sites or roads.

Contains a polygon of the area of the Landsat scene with path 224 and row 63 of the WRS2 system.
}
\details{
the files are
lsat_extent_LL: landsat extent for path 224 r63 (geographic coordinates lat long)
pa: protected areas of the area covered by lsat_extent (projection UTM north WGS84)
roads: roads of the area covered by lsat_extent (projection UTM north WGS84)
}
\examples{
load_vector()
plot(lsat_extent_LL)
##
plot(pa)
plot(abspres_poi,col="yellow",add=T)
}
\author{
Martin Wegmann
}
\references{
\href{http://remote-sensing-biodiversity.org/rsecology-book}.
}
\seealso{
load_raster()
}
\keyword{data}
\keyword{frames}
\keyword{shape}
\keyword{spatial}
\keyword{vector,}

